+++ rollup.externals.js
@@ -1,5 +1,6 @@
 const externals = [
   'assert',
+  'axios',
   'classnames',
   'lodash',
   'prop-types',
@@ -8,6 +9,7 @@ const externals = [
   'wix-rich-content-editor-common',
   'wix-rich-content-common',
   'react-i18next',
+  'react-flip-move',
   /^punycode$/,
   /^@wix\/draft-js$/,
   /^jss$/, //issue with ESM in CJS